Use fake database in tests
